extract populate() function
